+++ packages/auth/src/Auth.ts
@@ -147,6 +147,7 @@ export default class AuthClass {
                     logger.debug("Cognito Hosted authentication result", result);
                     that.currentSession().then(async (session) => {
                         try {
+                            await Credentials.clear();
                             const cred = await Credentials.set(session, 'session');
                             logger.debug('sign in succefully with', cred);
                         } catch (e) {
@@ -293,6 +294,7 @@ export default class AuthClass {
                 delete(user['challengeName']);
                 delete(user['challengeParam']);
                 try {
+                    await Credentials.clear();
                     const cred = await Credentials.set(session, 'session');
                     logger.debug('succeed to get cognito credentials', cred);
                 } catch (e) {
@@ -570,6 +572,7 @@ export default class AuthClass {
                     onSuccess: async (session) => {
                         logger.debug(session);
                         try {
+                            await Credentials.clear();
                             const cred = await Credentials.set(session, 'session');
                             logger.debug('succeed to get cognito credentials', cred);
                         } catch (e) {
@@ -602,6 +605,7 @@ export default class AuthClass {
                 onSuccess: async (session) => {
                     logger.debug(session);
                     try {
+                        await Credentials.clear();
                         const cred = await Credentials.set(session, 'session');
                         logger.debug('succeed to get cognito credentials', cred);
                     } catch (e) {