+++ packages/aws-amplify-react/src/Interactions/ChatBot.jsx
@@ -4,6 +4,7 @@ import { Input, Button } from "../AmplifyTheme";
 
 import { I18n } from '@aws-amplify/core';
 import { Interactions } from '../Categories';
+import regeneratorRuntime from 'regenerator-runtime/runtime';
 
 const styles = {
     itemMe: {