+++ packages/interactions/src/index.ts
@@ -29,3 +29,5 @@ Amplify.register(Interactions);
 export default Interactions;
 
 export * from './Providers/AWSLexProvider';
+
+export { InteractionsClass };