+++ packages/aws-amplify-react-native/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.20-unstable.1"></a>
+## [0.2.20-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.20-unstable.0...aws-amplify-react-native@0.2.20-unstable.1) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.20-unstable.0"></a>
+## [0.2.20-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19...aws-amplify-react-native@0.2.20-unstable.0) (2018-07-02)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19"></a>
+## [0.2.19](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.5...aws-amplify-react-native@0.2.19) (2018-06-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19-unstable.5"></a>
+## [0.2.19-unstable.5](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.4...aws-amplify-react-native@0.2.19-unstable.5) (2018-06-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19-unstable.4"></a>
+## [0.2.19-unstable.4](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.3...aws-amplify-react-native@0.2.19-unstable.4) (2018-06-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
 <a name="0.2.19-unstable.3"></a>
 ## [0.2.19-unstable.3](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.2...aws-amplify-react-native@0.2.19-unstable.3) (2018-06-28)